#138: extract and and complete performances store
