+++ src/components/conference/NewConference/Page/ButtonFooterRow.svelte
@@ -9,7 +9,7 @@
         endDate,
         intervalWarning,
         pending,
-    } from "../stores.ts";
+    } from "../stores/stores.ts";
     import darkTheme from "../../../../stores/theme-store";
     import Spinner from "../Spinner.svelte";
