+++ src/routes/konferanser/+page.svelte
@@ -8,7 +8,7 @@
 	import NewConferenceModal from '../../components/conference/NewConference/Modal/NewConferenceModal.svelte';
 	import type { IConferencesPageLoadData } from './+page.server';
 	import type { ConferenceCategoryType } from '../../enums/conference-category';
-    import {displayNewConferenceModal, initStore} from '../../components/conference/NewConference/stores';
+    import {displayNewConferenceModal, initStore} from '../../components/conference/NewConference/stores/stores';
 	import ConferenceCategoryCheckboxGroup from '../../components/checkbox/conference-category-checkbox-group/ConferenceCategoryCheckboxGroup.svelte';
 	import FilterConferenceCategoryModal from '../../components/modal/filter-conference-category-modal/FilterConferenceCategoryModal.svelte';
 	import { goto } from '$app/navigation';
